// src/app/model/prospect-direct.ts
export interface ProspectAppelDirect {

    id: number;
    civilite : string;
    nom: string;
    telephone: string;
    isChaudiere: boolean;
    isSolaire: boolean;
    isDivers: boolean;
    isIsolation: boolean;
    isAppel: boolean;
    isMessagerie: boolean;
    isRdv: boolean;
    isPasInteresse: boolean;




}